renamed all-teams-info to general-info
