+++ src/routes/routes.js
@@ -4,7 +4,7 @@ import {BrowserRouter, Route, Switch, Redirect} from 'react-router-dom'
 import ScoutingForm from '../views/scouting-form'
 import TeamInfo from '../views/team-info'
 import InfoHub from '../views/info-hub'
-import GeneralInfo from '../views/all-teams-info'
+import GeneralInfo from '../views/general-info'
 
 /**
  * This module is responsible for displaying different components depending on the url (emulating a multi page app)
