+++ src/views/general-info.js
